Working on database
Made a plan for finishing webscraping for now. Slight modification to merge now not rejecting if title is same only if link is same. Added version 2 of scraped Data which is around a third of the size. Going from 68k lines to 21k.

The issue is few special characters like / - ' are represented in unicdoe and look like \uxxx

I am not exactly sure how to fix it i tried for a bit and nothing seemed to work. There is probably an easy fix. Also I dont think this hsould be too much of a problem